work/taka-workspace/shared-project/django-project/myproject/requirements.txt
Django>=3.2,<4
psycopg2-binary>=2.9.1
djangorestframework>=3.13.0
django-filter>=0.11.0
requests~=2.28.2
python-decouple~=3.5
